pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "dotfile_manager"
version = "0.4.0"
authors = [
  { name="Lionel Gulich", email="lgulich@ethz.ch" },
]
description = "A tool for managing dotfiles."
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "PyYAML",
    "typeguard",
]

[project.urls]
"Homepage" = "https://github.com/lgulich/dotfile_manager"

[project.scripts]
dotfile_manager = "dotfile_manager.dotfile_manager:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["dotfile_manager*"]
exclude = ["test*"]
